Fix bugs and reduce code duplication
- Fix broken search on spirits page (wrong input ID and typo in placeholder)
- Add null check for search input to prevent JS errors
- Add missing </section> tag in classics.html
- Remove console.log from production code
- Add aria-label to all search buttons for accessibility
- Remove duplicate CSS mobile styles
- Extract shared format_json_file() to new utils.py module
- Remove unused os imports from utility scripts
