+++ docs/seasonal.html
@@ -30,7 +30,7 @@
       <div class="search-container">
         <div class="search-bar">
           <input type="text" class="search-input" id="cocktail-search" placeholder="Search cocktails...">
-          <button class="search-btn"><i class="fas fa-search"></i></button>
+          <button class="search-btn" aria-label="Search"><i class="fas fa-search"></i></button>
         </div>
       </div>
