+++ docs/spirits.html
@@ -27,8 +27,8 @@
       
       <div class="search-container">
         <div class="search-bar">
-          <input type="text" class="search-input" id="cocktail-search" placeholder="Search spiritsc...">
-          <button class="search-btn"><i class="fas fa-search"></i></button>
+          <input type="text" class="search-input" id="spirit-search" placeholder="Search spirits...">
+          <button class="search-btn" aria-label="Search"><i class="fas fa-search"></i></button>
         </div>
       </div>
       
@@ -278,19 +278,21 @@
       });
       
       // Search functionality
-      searchInput.addEventListener('input', () => {
+      if (searchInput) {
+        searchInput.addEventListener('input', () => {
         const searchTerm = searchInput.value.toLowerCase();
         
         beverages.forEach(beverage => {
           const beverageText = beverage.textContent.toLowerCase();
-          
+
           if (beverageText.includes(searchTerm)) {
             beverage.style.display = 'block';
           } else {
             beverage.style.display = 'none';
           }
         });
-      });
+        });
+      }
     });
   </script>
 </body>
