+++ docs/style.css
@@ -562,45 +562,8 @@ footer {
     font-size: 2em;
   }
 
-  .filter-controls {
-    flex-direction: column;
-    align-items: center;
-  }
-
   .hero-logo {
     max-width: 90%;
-  }
-  
-  /* More compact header for mobile */
-  header#main-header nav {
-    padding: 0.25rem 0;
-    overflow-x: auto;
-    -webkit-overflow-scrolling: touch;
-    white-space: nowrap;
-    display: flex;
-    justify-content: center;
-    margin-top: 0;
-    width: 100%;
-  }
-
-  
-  /* Make the logo smaller */
-  .seal-logo {
-    max-width: 60px;
-    max-height: 35px;
-    margin: 0.25rem auto;
-  }
-  
-  /* Optimize navigation for mobile */
-  nav {
-    margin-top: 0.25rem;
-    padding: 0;
-  }
-  
-  nav a {
-    margin: 0 0.5rem;
-    font-size: 0.8rem;
-    padding: 0.25rem 0;
   }
 }
